docs: step4 ResultOfCar dto생성
