+++ src/main/java/step4/domain/GameRound.java
@@ -1,6 +1,7 @@
 package step4.domain;
 
 import step4.domain.cargameStrategy.CarRacingGameStrategy;
+import step4.domain.dto.ResultOfCar;
 import step4.exception.RacingGameException;
 
 import java.util.ArrayList;
@@ -31,13 +32,13 @@ public class GameRound {
 
     private void decideWinner() {
         int max = gameRoundResult.stream()
-                .mapToInt(r -> r.position)
+                .mapToInt(ResultOfCar::getPosition)
                 .max()
                 .orElseThrow(() -> new RacingGameException("winner가 없습니다"));
 
         winner = gameRoundResult.stream()
-                .filter(r -> (r.position == max))
-                .map(r -> r.carName)
+                .filter(r -> (r.getPosition() == max))
+                .map(ResultOfCar::getCarName)
                 .collect(Collectors.toList());
 
     }
@@ -51,27 +52,11 @@ public class GameRound {
     }
 
     public List<ResultOfCar> getGameRoundResult() {
-        if(gameRoundResult==null){
+        if (gameRoundResult == null) {
             throw new RacingGameException("아직 게임을 진행하지 않았습니다.");
         }
         return Collections.unmodifiableList(gameRoundResult);
     }
 
-    public static class ResultOfCar {
-        private String carName;
-        private int position;
 
-        public ResultOfCar(String carName, int position) {
-            this.carName = carName;
-            this.position = position;
-        }
-
-        public String getCarName() {
-            return carName;
-        }
-
-        public int getPositoin() {
-            return position;
-        }
-    }
 }
